schedule関連のapiエンドポイントの見直し
